neurdicom/requirements.txt
absl-py==0.1.10
bleach==1.5.0
cycler==0.10.0
dipy==0.13.0
Django==2.0.2
djangorestframework==3.7.7
h5py==2.7.1
html5lib==0.9999999
Markdown==2.6.11
matplotlib==2.1.2
nibabel==2.2.1
numpy==1.14.0
protobuf==3.5.1
pydicom==1.0.0
pynetdicom3==0.1.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
Werkzeug==0.14.1
